Incorproate `preact` model, genericize `VTree`.
This adds patching (2-phase diffing) to miso. This allows patch objects to be
transferred elsewhere for rendering (in the case of miso native
this is required). This includes transferring event listener initialization,
component mounting / unmounting, and render tree diffing.

This patch abstract `VTree` over the underlying `DOMRef`. In the Haskell
runtime it will be represented as a pure identifier (`NodeId` which
holds the pure identifier `number`).

Patch application includes an `Environment` that holds a `Node`
mapping for each `Component`. This mapping is used to faciliate
modifications to the tree in O(1) time.

- [x] Abstract `VTree` and `Context` over `T` in `dom.ts`, tests, etc.
- [x] Add `patch.ts` and additional types for runtime patching
- [ ] Add `nodeId` to each `VTree` node
- [ ] Test patch accumulation and patch application for `applyPatch`
- [ ] Ensure `NodeId` can be used with event delegation